修改2个bug； 1.没有选择抄送人时产生空指针异常 2.菜品下单总金额错误，食堂订单菜品数量只能为个位数。

+++ app/src/main/java/com/haokuo/wenyanoa/network/bean/BuyFoodInBasketParams.java
@@ -9,10 +9,10 @@ import lombok.Data;
  */
 @Data
 public class BuyFoodInBasketParams extends UserIdApiKeyParams {
-    private int total;
+    private double total;
     private String ids;
 
-    public BuyFoodInBasketParams(int userId, String apiKey, int total, String ids) {
+    public BuyFoodInBasketParams(int userId, String apiKey, double total, String ids) {
         super(userId, apiKey);
         this.total = total;
         this.ids = ids;
